Refactor code with bamm samm
