--- sonar-project.properties
@@ -1,2 +1,0 @@
-sonar.sources = backend/src/
-sonar.cpd.exclusions=**/SDKAccessHelperBAMM.java
